refactor: Use comm component for history's and overview's type field.
