+++ packages/neuron-ui/src/components/FormattedTokenAmount/index.tsx
@@ -8,25 +8,30 @@ import { HIDE_BALANCE } from 'utils/const'
 
 import styles from './formattedTokenAmount.module.scss'
 
-type FormattedTokenAmountProps = { item: State.Transaction; show: boolean }
+type FormattedTokenAmountProps = { item: State.Transaction; show: boolean; symbolClassName?: string }
 type AmountProps = Omit<FormattedTokenAmountProps, 'isNeedCopy'> & {
   sudtAmount: string
   isReceive: boolean
   amount: string
+  symbolClassName?: string
 }
 
-const Amount = ({ sudtAmount, show, item, isReceive, amount }: AmountProps) => {
+const Amount = ({ sudtAmount, show, item, isReceive, amount, symbolClassName }: AmountProps) => {
   return sudtAmount ? (
     <div className={show && !sudtAmount.includes('-') ? styles.isReceive : ''}>
       {show ? `${!sudtAmount.includes('-') ? '+' : ''}${sudtAmount}` : HIDE_BALANCE}&nbsp;
-      <UANTonkenSymbol name={item.sudtInfo!.sUDT.tokenName} symbol={item.sudtInfo!.sUDT.symbol} />
+      <UANTonkenSymbol
+        className={symbolClassName}
+        name={item.sudtInfo!.sUDT.tokenName}
+        symbol={item.sudtInfo!.sUDT.symbol}
+      />
     </div>
   ) : (
     <span className={show && isReceive ? styles.isReceive : ''}>{amount}</span>
   )
 }
 
-export const FormattedTokenAmount = ({ item, show }: FormattedTokenAmountProps) => {
+export const FormattedTokenAmount = ({ item, show, symbolClassName }: FormattedTokenAmountProps) => {
   let amount = '--'
   let sudtAmount = ''
   let copyText = amount
@@ -51,7 +56,7 @@ export const FormattedTokenAmount = ({ item, show }: FormattedTokenAmountProps) 
     }
   }
 
-  const props = { sudtAmount, show, item, isReceive, amount }
+  const props = { sudtAmount, show, item, isReceive, amount, symbolClassName }
 
   return show ? (
     <CopyZone content={copyText}>
