+++ packages/neuron-ui/src/components/Overview/index.tsx
@@ -1,11 +1,10 @@
 import React, { useCallback, useMemo, useEffect, useState } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
-import { Trans, useTranslation } from 'react-i18next'
+import { useTranslation } from 'react-i18next'
 import { useState as useGlobalState, useDispatch, updateTransactionList } from 'states'
 
 import { shannonToCKBFormatter, uniformTimeFormatter, backToTop, CONSTANTS, RoutePath, useFirstLoadWallet } from 'utils'
 
-import { UANTokenName } from 'components/UANDisplay'
 import PageContainer from 'components/PageContainer'
 import TransactionStatusWrap from 'components/TransactionStatusWrap'
 import FormattedTokenAmount from 'components/FormattedTokenAmount'
@@ -17,38 +16,10 @@ import { ArrowNext, EyesClose, EyesOpen, OverviewSend, OverviewReceive, Addressb
 import BalanceSyncIcon from 'components/BalanceSyncingIcon'
 import CopyZone from 'widgets/CopyZone'
 import { HIDE_BALANCE } from 'utils/const'
+import TracsactionType from 'components/TransactionType'
 import styles from './overview.module.scss'
 
 const { PAGE_SIZE, CONFIRMATION_THRESHOLD } = CONSTANTS
-
-const genTypeLabel = (
-  type: 'send' | 'receive' | 'create' | 'destroy',
-  status: 'pending' | 'confirming' | 'success' | 'failed'
-) => {
-  switch (type) {
-    case 'send': {
-      if (status === 'failed') {
-        return 'sent'
-      }
-      if (status === 'pending' || status === 'confirming') {
-        return 'sending'
-      }
-      return 'sent'
-    }
-    case 'receive': {
-      if (status === 'failed') {
-        return 'received'
-      }
-      if (status === 'pending' || status === 'confirming') {
-        return 'receiving'
-      }
-      return 'received'
-    }
-    default: {
-      return type
-    }
-  }
-}
 
 const TransactionStatus = ({
   item,
@@ -76,72 +47,6 @@ const TransactionStatus = ({
       <TransactionStatusWrap status={status} confirmationCount={confirmationCount} />
     </div>
   )
-}
-
-const TracsactionType = ({
-  item,
-  cacheTipBlockNumber,
-  bestKnownBlockNumber,
-}: {
-  item: Omit<State.Transaction, 'status'> & { status: State.Transaction['status'] | 'confirming' }
-  cacheTipBlockNumber: number
-  bestKnownBlockNumber: number
-}) => {
-  const [t] = useTranslation()
-  let typeLabel: string = '--'
-  let { status } = item
-  let typeTransProps: {
-    i18nKey: string
-    components: JSX.Element[]
-  } = {
-    i18nKey: '',
-    components: [],
-  }
-
-  if (item.blockNumber !== undefined) {
-    const confirmationCount =
-      item.blockNumber === null || item.status === 'failed'
-        ? 0
-        : 1 + Math.max(cacheTipBlockNumber, bestKnownBlockNumber) - +item.blockNumber
-
-    if (status === 'success' && confirmationCount < CONFIRMATION_THRESHOLD) {
-      status = 'confirming'
-    }
-    if (item.nftInfo) {
-      // NFT
-      const { type } = item.nftInfo
-      typeLabel = `${t(`overview.${genTypeLabel(type, status)}`)}`
-    } else if (item.sudtInfo?.sUDT) {
-      // Asset Account
-      if (['create', 'destroy'].includes(item.type)) {
-        // create/destroy an account
-        typeTransProps = {
-          i18nKey: `overview.${item.type}SUDT`,
-          components: [
-            <UANTokenName
-              name={item.sudtInfo.sUDT.tokenName}
-              symbol={item.sudtInfo.sUDT.symbol}
-              className={styles.tokenName}
-            />,
-          ],
-        }
-      } else {
-        // send/receive to/from an account
-        const type = +item.sudtInfo.amount <= 0 ? 'send' : 'receive'
-        typeLabel = `UDT ${t(`overview.${genTypeLabel(type, status)}`)}`
-      }
-    } else if (item.type === 'create' || item.type === 'destroy') {
-      // normal tx
-      if (item.assetAccountType === 'CKB') {
-        typeLabel = `${t(`overview.${item.type}`, { name: 'CKB' })}`
-      } else {
-        typeLabel = `${t(`overview.${item.type}`, { name: 'Unknown' })}`
-      }
-    } else {
-      typeLabel = item.nervosDao ? 'Nervos DAO' : t(`overview.${genTypeLabel(item.type, status)}`)
-    }
-  }
-  return typeTransProps.i18nKey ? <Trans {...typeTransProps} /> : <>{typeLabel}</>
 }
 
 const Overview = () => {
